+++ FinancialStructures/Database/Statistics/Implementation/Statistic.RecentChange.cs
@@ -13,14 +13,14 @@ namespace FinancialStructures.Database.Statistics.Implementation
         /// <inheritdoc/>
         public override void Calculate(IPortfolio portfolio, Account account, TwoName name)
         {
-            Value = portfolio.RecentChange(account, name);
+            Value = (double)portfolio.RecentChange(account, name);
             fCurrency = portfolio.BaseCurrency;
         }
 
         /// <inheritdoc/>
         public override void Calculate(IPortfolio portfolio, Totals total, TwoName name)
         {
-            Value = portfolio.RecentChange(total, name);
+            Value = (double)portfolio.RecentChange(total, name);
             fCurrency = portfolio.BaseCurrency;
         }
     }
