+++ FinancialStructures/FinanceStructures/IValueList.cs
@@ -107,7 +107,7 @@ namespace FinancialStructures.FinanceStructures
         /// <param name="value">The value data to add.</param>
         /// <param name="reportLogger">An optional logger to log progress.</param>
         /// <returns>Was adding or editing successful.</returns>
-        bool TryEditData(DateTime oldDate, DateTime date, double value, IReportLogger reportLogger = null);
+        bool TryEditData(DateTime oldDate, DateTime date, decimal value, IReportLogger reportLogger = null);
 
         /// <summary>
         /// Sets data on the date specified to the value given. This overwrites the existing
@@ -116,7 +116,7 @@ namespace FinancialStructures.FinanceStructures
         /// <param name="date">The date to add data to.</param>
         /// <param name="value">The value data to add.</param>
         /// <param name="reportLogger">An optional logger to log progress.</param>
-        void SetData(DateTime date, double value, IReportLogger reportLogger = null);
+        void SetData(DateTime date, decimal value, IReportLogger reportLogger = null);
 
         /// <summary>
         /// Attempts to delete data on the date specified.
